+++ Services/BeforeTheScholarship.Services.DebtService/Models/UpdateDebtModel.cs
@@ -41,7 +41,7 @@ public class UpdateDebtModelValidator : AbstractValidator<UpdateDebtModel>
         if (string.IsNullOrEmpty(number))
             return true;
 
-        return int.TryParse(number, out _);
+        return long.TryParse(number, out _);
     }
 }
